+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 4.6.0
 
+### Features
+
+- [#585](https://github.com/okta/okta-auth-js/pull/585) Uses native fetch, if available
+
 ### Other
 
 - [#583](https://github.com/okta/okta-auth-js/pull/583) Better error handling for redirect flows: if redirect URI contains `error` or `error_description` then `isLoginRedirect` will return true and `parseFromUrl` will throw `OAuthError`